+++ RazorSvelte/App/modal.svelte
@@ -1,0 +1,16 @@
+<script lang="ts">
+    import Layout from "./shared/layout.svelte";
+</script>
+
+<Layout>
+    <div class="container text-center">
+        <h1>Modal</h1>
+        
+        <p>
+            Content here
+        </p>
+    </div>
+</Layout>
+
+<style lang="scss">
+</style>