+++ RazorSvelte/App/shared/layout.svelte
@@ -2,9 +2,11 @@
     import Layout from "../lib/offcanvas-layout.svelte";
     //import Layout from "../lib/header-layout.svelte";
     import Links from "./link-list-items.svelte";
+
+    export let title: string | undefined = undefined;
 </script>
 
-<Layout>
+<Layout title={title}>
     <svelte:fragment slot="links"><Links /></svelte:fragment>
     <slot></slot>
 </Layout>