+++ RazorSvelte/Endpoints/CountryEndpoints.cs
@@ -1,4 +1,6 @@
-﻿using RazorSvelte.SampleData;
+﻿using System.Xml.Linq;
+using Microsoft.AspNetCore.Mvc;
+using RazorSvelte.SampleData;
 
 namespace RazorSvelte.Endpoints;
 
@@ -8,10 +10,22 @@ public class CountryEndpoints
 
     public static void UseEndpoints(WebApplication app)
     {
-        app.MapGet(CountriesUrl, () =>
+        app.MapGet(CountriesUrl, (
+            [FromQuery(Name = "culture-contains")] 
+            string? cultureContains,
+            [FromQuery] int? limit) =>
         {
             var json = File.ReadAllText("./SampleData/countries.json");
-            return json.Deserialize<Country[]>();
+
+            return json?.Deserialize<Country[]>()?
+                .Where(c => string.IsNullOrEmpty(cultureContains) || c?.Culture?.Contains(cultureContains, StringComparison.OrdinalIgnoreCase) == true)
+                .Select(c => 
+                {
+                    c.Timestamp = DateTime.Now;
+                    return c;
+                })
+                .OrderByDescending(c => c.Code)
+                .Take(limit ?? int.MaxValue);
         });
     }
 }