+++ RazorSvelte/Pages/Modal.cshtml.cs
@@ -1,0 +1,7 @@
+namespace RazorSvelte.Pages;
+
+public partial class Urls
+{
+    public const string  ModalUrl = "/modal";
+}
+public class ModalModel : PageModel {}