+++ RazorSvelte/Scripts/build-all.js
@@ -1,6 +1,14 @@
 ﻿const fs = require("fs");
 const path = require("path");
 const cp = require('child_process');
+const config = require(`./config`);
+
+const root = config.root;
+const build = config.build;
+const pagesDir = config.pagesDir;
+const rollupExt = config.rollupExt;
+const mapExt = config.mapExt;
+const rollupExtraArgs = config.rollupExtraArgs;
 
 const exec = cmd => new Promise(resolve => {
     console.log(cmd);
@@ -18,7 +26,7 @@ const getAllConfigs = function (dir, result) {
         if (fs.statSync(path.join(dir, file)).isDirectory()) {
             result = getAllConfigs(path.join(dir, file), result)
         } else {
-            if (file.endsWith(".rollup.config.js")) {
+            if (file.endsWith(rollupExt)) {
                 result.push("./" + path.join(dir, file).replace(/[\\/]/g, "/"));
             }
         }
@@ -29,7 +37,7 @@ const getAllConfigs = function (dir, result) {
 
 const removeMaps = root => new Promise(resolve => {
     for (let file of fs.readdirSync(root)) {
-        if (file.endsWith(".map")) {
+        if (file.endsWith(mapExt)) {
             var name = (root + "/" + file).replace("//", "/");
             console.log("removing extra file: ", name);
             try {
@@ -44,34 +52,21 @@ const removeMaps = root => new Promise(resolve => {
 });
 
 const promises = [];
-const root = "./wwwroot/";
-const build = "./wwwroot/build/";
 
 if (!fs.existsSync(build)) {
     console.log("Creating dir " + build + " ...");
     fs.mkdirSync(build);
 } else {
-    for (let file of fs.readdirSync(build)) {
-        if (file.endsWith(".map")) {
-            var name = (build + "/" + file).replace("//", "/");
-            console.log("removing ", name);
-            try {
-                fs.unlinkSync(name)
-                //file removed
-            } catch (err) {
-                console.error(err)
-            }
-        }
-    }
+    removeMaps(build);
 }
 
 promises.push(exec(`npm run fe-scss-build`));
 
-for (let config of getAllConfigs("./Pages")) {
-    promises.push(exec("npx rollup -c " + config + " --bundleConfigAsCjs"));
+for (let config of getAllConfigs(pagesDir)) {
+    promises.push(exec("npx rollup -c " + config + rollupExtraArgs));
 }
 
 promises.push(removeMaps(root));
 promises.push(removeMaps(build));
 
-Promise.all(promises).then(() => console.log("build all done!"));
+Promise.all(promises).then(() => console.log("Build all done!"));